Add time-to-live limit for persisted log entries.
If there is a programmer error, where a non-retryable error is treated
as a retryable error, the entry would never be deleted from disk and
would be retried each time the app restarts.

This patch adds a limit based on the number of seconds after
entry creation. After which the entry is dropped.

Also add tests.